% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuser.r
\name{expand_ct}
\alias{expand_ct}
\title{Detect & Validate Syracuse, NY Census Tract Format Variants}
\usage{
expand_ct(tracts, order = TRUE, duplicate.rm = FALSE, na.rm = TRUE)
}
\arguments{
\item{tracts}{A series of census tracts of class \code{vector}, either as
an independent object or an extraction from a \code{data.frame} object in
\code{data$variable} syntax.}

\item{order}{Logical \code{TRUE} or \code{FALSE} indicating whether output
should be in numeric or alphanumeric order. Defaults to \code{TRUE}.}

\item{duplicate.rm}{Logical \code{TRUE} or \code{FALSE} indicating whether
output should remove duplicate census tract values. Defaults to \code{FALSE}.}

\item{na.rm}{Logical \code{TRUE} or \code{FALSE} indicating whether output
should remove incomplete reformatting due to missing (\code{NA}) values.
Defaults to \code{TRUE}.}
}
\value{
A tabular object of class \code{data.frame} containing five variants
of 55 Syracuse, NY census tracts. Variants may be previewed in package
\code{cuser} dataset \code{tracts}. Output may be ordered with duplicates
and missing values removed depending on inputs to logical arguments
\code{order =}, \code{duplicate.rm =}, and \code{na.rm = }, respectively.
}
\description{
\code{expand_ct()} accepts a series of census tract values of class
\code{vector}, detects their format, and returns an object of
class \code{data.frame} containing four other variants for speedy
reformatting. Sorting is optional, as is preservation of duplicates
and missing values.
}
\seealso{
\code{tracts}, \code{merge}, \code{order}, \code{duplicated}, \code{complete.cases}
}
\author{
Jamison R. Crawford, Social Sector Data Science Consultant
}
