% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuser.r
\name{filter_ta}
\alias{filter_ta}
\title{Clean, Filter, & Order DSS Temporary Assistance Data}
\usage{
filter_ta(data, tracts = 1, cases = 2, individuals = 3)
}
\arguments{
\item{data}{A tabular object of class \code{data.frame} containing
DSS TA data.}

\item{tracts}{The column number of the \code{data.frame} object passed to
argument \code{data =} which contains the table's census tract values.
Defaults to \code{1}.}

\item{cases}{The column number of the \code{data.frame} object passed to
argument \code{data =} which contains the table's TA cases. Defaults to
\code{2}.}

\item{individuals}{The column number of the \code{data.frame} object passed
to argument \code{data =} which contains the table's TA-assisted
individuals. Defaults to \code{3}.}
}
\value{
A tabular object of class \code{data.frame} containing only DSS TA
data pertaining to the 55 census tracts in Syracuse, NY, including tracts in
which no TA cases are reported. Arguments \code{tracts =}, \code{cases =},
and \code{individuals} are renamed \code{tract}, \code{cases}, and
\code{individuals}, prefaced by the FIPS values in column \code{geoid}.
}
\description{
\code{filter_ta()} processes tables of class \code{data.frame}
by filtering data exclusively to the 55 census tracts within Syracuse,
NY, ordering values according to Federal Information Processing Standard
Publication 6-4 (FIPS) GEOID, and preserving census tracts which exclude
Temporary Assistance (TA) data that are not reported by Onondaga County
Department of Social Services (DSS).
}
\seealso{
\code{merge}, \code{order}
}
\author{
Jamison R. Crawford, Social Sector Data Science Consultant
}
